requirements.txt
beautifulsoup4~=4.12
lxml~=5.0
python-dotenv~=1.0

# development / testing
pytest~=8.0
